// src/models/schemas/RefreshToken.schema.ts
import { ObjectId } from 'mongodb'

interface RefreshTokenType {
  _id?: ObjectId
  token: string
  create_at?: Date
  user_id: ObjectId
  iat?: number
  exp?: number
}

export default class RefreshToken {
  _id?: ObjectId
  token: string
  create_at: Date
  user_id: ObjectId
  iat?: number
  exp?: number
  constructor({ _id, token, create_at, user_id, iat, exp }: RefreshTokenType) {
    this._id = _id
    this.token = token
    this.create_at = create_at || new Date()
    this.user_id = user_id
    this.iat = iat
    this.exp = exp
  }
}